Send Host header with OCSP request
Older versions of openssl do not send this header by themselves. But
there is an undocumented option to add headers to the request. If the
OCSP server requires a Host header, the requests will fail with these
old versions of openssl. This adds the Host header to the request.

This can be dropped once only versions of openssl >= 1.1.0 should be
supported. These version send the Host header by default.